Move Storage, Parser and Ui into utils
