+++ text-ui-test/EXPECTED.TXT
@@ -1,13 +1,13 @@
 So you require my assistance?
 Understood, I have added to your list:
 [T][ ] CS2103
-You have 1 screwllum.Parser.tasks
+You have 1 screwllum.utils.Parser.tasks
 Understood, I have added to your list:
 [D][ ] life  (by: 2024)
-You have 2 screwllum.Parser.tasks
+You have 2 screwllum.utils.Parser.tasks
 Understood, I have added to your list:
 [E][ ] Birthday (from: 1990 to: 2000)
-You have 3 screwllum.Parser.tasks
+You have 3 screwllum.utils.Parser.tasks
 Let me help you to remember what you need to do ...
 1. [T][ ] CS2103
 2. [D][ ] life  (by: 2024)
